requirements.txt
Django==1.10.2
djangorestframework==3.5.1
django-oauth-toolkit==0.10.0
django-extensions==1.7.4
psycopg2==2.6.2
